shared droppdown working and much faster
