+++ src/app/createcontest/createcontest.component.scss
@@ -10,3 +10,17 @@ table{
 mat-radio-group.mat-input-invalid{
     border-bottom: 1px solid #f44336;
 }
+.dropdown-form {
+  min-width: 150px;
+  max-width: 500px;
+  width: 100%;
+}
+.dropdown-full-width {
+    padding: 10px;
+  width: 100%;
+}
+
+.mat-active{
+    background-color: #bbb;
+}
+
