--- src/app/createcontest/team-dropdown/team-dropdown.component.html
@@ -1,12 +1,0 @@
-<mat-form-field class="dropdown-full-width" [ngClass]="fieldCss()" [formGroup]="teamCtrlGroup">
-    <input matInput placeholder="{{seed}} {{region}}" aria-label="Team"
-           [matAutocomplete]="auto" formControlName="teamFormCtrl">
-    <mat-autocomplete #auto="matAutocomplete" autoActiveFirstOption >
-      <mat-option *ngFor="let team of filteredTeams| async" [value]="team.name">
-        <!--<img style="vertical-align:middle;" aria-hidden
-        src= https://sports.cbsimg.net/images/collegebasketball/logos/50x50/{{team.short}}.png" height="25" />-->
-        <small>{{ team.detail }}</small>
-      </mat-option>
-    </mat-autocomplete>
-</mat-form-field>
-
