--- src/app/createcontest/team-dropdown/team-dropdown.component.scss
@@ -1,14 +1,0 @@
-.dropdown-form {
-  min-width: 150px;
-  max-width: 500px;
-  width: 100%;
-}
-.dropdown-full-width {
-    padding: 10px;
-  width: 100%;
-}
-
-.mat-active{
-    background-color: #bbb;
-}
-
